+++ src/__tests__/controller.test.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import {
+  act as actComponent,
   fireEvent,
   render,
   screen,
@@ -8,11 +9,12 @@ import {
 } from '@testing-library/react';
 
 import { Controller } from '../controller';
-import { ControllerRenderProps, FieldValues, WatchedForm } from '../types';
+import { ControllerRenderProps, FieldValues, ValidateResult } from '../types';
 import { useFieldArray } from '../useFieldArray';
 import { useForm } from '../useForm';
 import { FormProvider } from '../useFormContext';
 import { useWatch } from '../useWatch';
+import noop from '../utils/noop';
 
 function Input<TFieldValues extends FieldValues>({
   onChange,
@@ -260,6 +262,61 @@ describe('Controller', () => {
     expect(touched).toEqual({ test: true });
   });
 
+  it('should set field to formState validatingFields and render field isValidating state', async () => {
+    jest.useFakeTimers();
+
+    const getValidateMock: (timeout: number) => Promise<ValidateResult> = (
+      timeout: number,
+    ) => {
+      return new Promise((resolve) => {
+        setTimeout(() => {
+          resolve(true);
+        }, timeout);
+      });
+    };
+
+    let validatingFields: any;
+    const Component = () => {
+      const { control, formState } = useForm({ mode: 'onBlur' });
+
+      validatingFields = formState.validatingFields;
+
+      return (
+        <Controller
+          defaultValue=""
+          name="test"
+          render={({ field, fieldState }) => (
+            <>
+              <div>isValidating: {String(fieldState.isValidating)}</div>
+              <input {...field} />
+            </>
+          )}
+          control={control}
+          rules={{
+            validate: () => getValidateMock(1000),
+          }}
+        />
+      );
+    };
+
+    render(<Component />);
+
+    expect(validatingFields).toEqual({});
+    expect(screen.getByText('isValidating: false')).toBeVisible();
+
+    fireEvent.blur(screen.getByRole('textbox'));
+
+    expect(validatingFields).toEqual({ test: true });
+    expect(screen.getByText('isValidating: true')).toBeVisible();
+
+    await actComponent(async () => {
+      jest.advanceTimersByTime(1100);
+    });
+
+    expect(validatingFields).toEqual({});
+    expect(screen.getByText('isValidating: false')).toBeVisible();
+  });
+
   it('should call trigger method when re-validate mode is onBlur with blur event', async () => {
     const Component = () => {
       const {
@@ -271,7 +328,7 @@ describe('Controller', () => {
       });
 
       return (
-        <form onSubmit={handleSubmit(() => {})}>
+        <form onSubmit={handleSubmit(noop)}>
           <Controller
             defaultValue=""
             name="test"
@@ -839,7 +896,7 @@ describe('Controller', () => {
   });
 
   it('should retain default value or defaultValues at Controller', () => {
-    let getValuesMethod = () => {};
+    let getValuesMethod = noop;
     const Component = () => {
       const { control, getValues } = useForm<{
         test: number;
@@ -982,7 +1039,7 @@ describe('Controller', () => {
     type FormValue = {
       test: string;
     };
-    const watchedValue: WatchedForm<FormValue>[] = [];
+    const watchedValue: FormValue[] = [];
     const Component = () => {
       const { control, watch } = useForm<FormValue>({
         defaultValues: {
@@ -1119,7 +1176,7 @@ describe('Controller', () => {
       });
 
       return (
-        <form onSubmit={handleSubmit(() => {})}>
+        <form onSubmit={handleSubmit(noop)}>
           {fields.map((field, index) => {
             return (
               <div key={field.id}>
@@ -1455,7 +1512,7 @@ describe('Controller', () => {
       const { control, handleSubmit } = useForm<{ numbers: number[] }>();
 
       return (
-        <form onSubmit={handleSubmit(() => {})}>
+        <form onSubmit={handleSubmit(noop)}>
           <Controller
             control={control}
             name="numbers"