+++ src/__tests__/useFieldArray.test.tsx
@@ -15,12 +15,12 @@ import {
   SubmitHandler,
   UseFormRegister,
   UseFormReturn,
-  WatchedForm,
 } from '../types';
 import { useFieldArray } from '../useFieldArray';
 import { useForm } from '../useForm';
 import { FormProvider } from '../useFormContext';
 import { useFormState } from '../useFormState';
+import noop from '../utils/noop';
 
 let i = 0;
 
@@ -2695,7 +2695,7 @@ describe('useFieldArray', () => {
       }[];
     };
 
-    const watchedValues: WatchedForm<FormValues>[] = [];
+    const watchedValues: FormValues[] = [];
 
     const Child = ({
       control,
@@ -2859,7 +2859,7 @@ describe('useFieldArray', () => {
 
       return (
         <React.StrictMode>
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             {fields.map((field, index) => {
               return (
                 <div key={field.id}>
@@ -3173,7 +3173,7 @@ describe('useFieldArray', () => {
   });
 
   it('should avoid omit keyName when defaultValues contains keyName attribute', () => {
-    let getValuesMethod: Function = () => {};
+    let getValuesMethod: Function = noop;
 
     const App = () => {
       const { control, getValues } = useForm({
@@ -3224,7 +3224,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             <p>{errors.test?.root?.message}</p>
             <button>submit</button>
             <button
@@ -3284,7 +3284,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             <p>{errors.test?.root?.message}</p>
             <button>submit</button>
             <button
@@ -3341,7 +3341,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             <p>{errors.test?.root?.message}</p>
             <button>submit</button>
             <button
@@ -3396,7 +3396,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             <p>{errors.test?.root?.message}</p>
             <button>submit</button>
             <button
@@ -3465,7 +3465,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             {fields.map((field, index) => {
               return (
                 <div key={field.id}>
@@ -3564,7 +3564,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             {fields.map((field, index) => {
               return (
                 <div key={field.id}>
@@ -3614,7 +3614,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             {fields.map((field, index) => {
               return (
                 <div key={field.id}>
@@ -3715,7 +3715,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             {fields.map((field, index) => {
               return (
                 <div key={field.id}>
@@ -3776,7 +3776,7 @@ describe('useFieldArray', () => {
         });
 
         return (
-          <form onSubmit={handleSubmit(() => {})}>
+          <form onSubmit={handleSubmit(noop)}>
             {fields.map((field, index) => {
               return (
                 <div key={field.id}>